requirements.txt
streamlit
pandas
requests
sqlalchemy
psycopg2-binary
python-dotenv
yfinance